refactor(settings): drop zustand from the overlay guard
The Escape-key count only needs to be read at the moment a key is pressed,
not rendered — a plain module counter does the job without a store,
subscriptions, or the isDisabled indirection through useKeyDown. AppSettings
now wraps `close` itself and checks the counter synchronously.
