+++ apps/client/src/common/components/dialog/Dialog.tsx
@@ -2,7 +2,7 @@ import { Dialog as BaseDialog } from '@base-ui/react/dialog';
 import type { ReactNode } from 'react';
 import { IoClose } from 'react-icons/io5';
 
-import { useRegisterOverlay } from '../../stores/overlayStore';
+import { useRegisterOverlay } from '../../utils/overlayRegistry';
 import IconButton from '../buttons/IconButton';
 
 import style from './Dialog.module.scss';
